src/python-backend/pyproject.toml
[tool.poetry]
name = "straysafe"
version = "0.1.0"
description = "Business logic of StaySafe web platform"
authors = [
    "Renaldas Narbutas <renaldas.narbutas@mif.stud.vu.lt>",
    "Arsenij Nikulin <arsenij.nikulin@mif.stud.vu.lt>",
    "Edvinas Gerdvila <edvinas.gerdvila@mif.stud.vu.lt>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.0"
opencv-python = "^4.9.0.80"
scikit-image = "^0.23.1"
psycopg2 = "^2.9.9"
matplotlib = "^3.9.0"

[tool.poetry.scripts]
straysafe = "straysafe.run:run_debug"
imgtest = "straysafe.run:img_test"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.3"
pytest = "^8.1.1"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
